fix(TUP-25391)[Proxy]Can't logon project after install Metadata Bridge successfully https://jira.talendforge.org/browse/TUP-25391
