+++ main/plugins/org.talend.core.runtime/src/main/java/org/talend/core/nexus/HttpClientTransport.java
@@ -35,8 +35,8 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.core.runtime.OperationCanceledException;
 import org.talend.commons.exception.BusinessException;
+import org.talend.commons.utils.network.IProxySelectorProvider;
 import org.talend.commons.utils.network.TalendProxySelector;
-import org.talend.commons.utils.network.TalendProxySelector.IProxySelectorProvider;
 import org.talend.core.runtime.CoreRuntimePlugin;
 import org.talend.core.runtime.maven.MavenArtifact;
 import org.talend.core.runtime.maven.MavenUrlHelper;
